requirements.txt
boto3>=1.34.0
sagemaker>=2.200.0
python-dotenv>=1.0.0
requests>=2.31.0
transformers>=4.35.0
torch>=2.1.0
sentencepiece>=0.1.99
datasets>=2.14.0
scikit-learn>=1.3.0
tqdm>=4.66.0
accelerate>=0.20.0
